Create stephanie4-8-21
Created the frame to go in each activity, added that to each of the activity functions as well as the crayons to each activity function. I also created the triangle tips for the crayons and am working on making them colored to the specific crayon color. I made "back", "next" and "?" buttons but for some reason Khan isn't letting me add/change buttons - I think it's being buggy right now so I will try again later.